# backend/src/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Endgame Grocery API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - ApiKeyAuth: []
paths:
  /lists:
    get:
      summary: List accessible grocery lists
      operationId: listGroceryLists
      responses:
        '200':
          description: Lists accessible to the API-key owner.
          content:
            application/json:
              schema:
                type: object
                required:
                  - lists
                properties:
                  lists:
                    type: array
                    items:
                      $ref: '#/components/schemas/List'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /lists/{listId}/items:
    parameters:
      - $ref: '#/components/parameters/ListId'
    get:
      summary: List grocery items
      operationId: listGroceryItems
      responses:
        '200':
          description: Items in the requested grocery list.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create a grocery item
      operationId: createGroceryItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
      responses:
        '201':
          description: The created item.
          content:
            application/json:
              schema:
                type: object
                required:
                  - item
                properties:
                  item:
                    $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /lists/{listId}/items/{itemId}/toggle:
    parameters:
      - $ref: '#/components/parameters/ListId'
      - $ref: '#/components/parameters/ItemId'
    post:
      summary: Toggle a grocery item status
      description: Toggles between `open` and `done`; toggles to `done` also update autocomplete history for the list.
      operationId: toggleGroceryItem
      responses:
        '200':
          description: The updated item.
          content:
            application/json:
              schema:
                type: object
                required:
                  - item
                properties:
                  item:
                    $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /lists/{listId}/items/{itemId}:
    parameters:
      - $ref: '#/components/parameters/ListId'
      - $ref: '#/components/parameters/ItemId'
    patch:
      summary: Rename a grocery item
      operationId: renameGroceryItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: The renamed item.
          content:
            application/json:
              schema:
                type: object
                required:
                  - item
                properties:
                  item:
                    $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a grocery item
      operationId: deleteGroceryItem
      responses:
        '204':
          description: Item deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    ListId:
      name: listId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Grocery list UUID.
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Grocery item UUID.
  schemas:
    List:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    Item:
      type: object
      required:
        - id
        - name
        - status
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - open
            - done
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: The request body is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: The API key is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The API-key owner cannot access the requested list.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The requested list or item was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'